refactor: move EF logic to another project and fix namespaces
