+++ Ziggurat.sln
@@ -18,6 +18,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{F0686707
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sample.Cap.Mongo", "samples\Sample.Cap.Mongo\Sample.Cap.Mongo.csproj", "{932BB534-9388-4847-9647-F14DDAAC290B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ziggurat.SqlServer", "src\Ziggurat.SqlServer\Ziggurat.SqlServer.csproj", "{77D691F4-BFC6-496F-9041-6F0344EAF680}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ziggurat.SqlServer.Tests", "tests\Ziggurat.SqlServer.Tests\Ziggurat.SqlServer.Tests.csproj", "{54E6B344-F154-469B-8E50-ACA05309A67A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -48,6 +52,14 @@ Global
 		{932BB534-9388-4847-9647-F14DDAAC290B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{932BB534-9388-4847-9647-F14DDAAC290B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{932BB534-9388-4847-9647-F14DDAAC290B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{77D691F4-BFC6-496F-9041-6F0344EAF680}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{77D691F4-BFC6-496F-9041-6F0344EAF680}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{77D691F4-BFC6-496F-9041-6F0344EAF680}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{77D691F4-BFC6-496F-9041-6F0344EAF680}.Release|Any CPU.Build.0 = Release|Any CPU
+		{54E6B344-F154-469B-8E50-ACA05309A67A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{54E6B344-F154-469B-8E50-ACA05309A67A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{54E6B344-F154-469B-8E50-ACA05309A67A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{54E6B344-F154-469B-8E50-ACA05309A67A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{5D6F3BFC-BD8B-45CA-BAB5-7D32ABB23C89} = {44743259-B342-4FFF-A6F4-30B0C2BFEB1B}
@@ -56,5 +68,7 @@ Global
 		{0350E066-110B-4323-970A-93B084BE09A3} = {69E60133-9529-4E6A-AF47-D51658BC97BD}
 		{2FDCE0BC-B5D7-41A6-BE40-DE2D707FE737} = {F0686707-3534-41C3-A71D-78C6CF11E7E6}
 		{932BB534-9388-4847-9647-F14DDAAC290B} = {44743259-B342-4FFF-A6F4-30B0C2BFEB1B}
+		{77D691F4-BFC6-496F-9041-6F0344EAF680} = {69E60133-9529-4E6A-AF47-D51658BC97BD}
+		{54E6B344-F154-469B-8E50-ACA05309A67A} = {F0686707-3534-41C3-A71D-78C6CF11E7E6}
 	EndGlobalSection
 EndGlobal
