+++ samples/Sample.Cap.Mongo/Consumer.cs
@@ -1,7 +1,5 @@
 ﻿using DotNetCore.CAP;
-using MongoDB.Driver;
 using Ziggurat;
-using Ziggurat.MongoDB;
 
 namespace Sample.Cap.Mongo;
 
@@ -18,29 +16,5 @@ public class Consumer : ICapSubscribe
     public async Task ConsumeMessage(MyMessage message)
     {
         await _service.ProcessMessageAsync(message);
-    }
-}
-
-public class ConsumerService : IConsumerService<MyMessage>
-{
-    private readonly ILogger<ConsumerService> _logger;
-    private readonly IMongoClient _client;
-
-    public ConsumerService(ILogger<ConsumerService> logger, IMongoClient client)
-    {
-        _logger = logger;
-        _client = client;
-    }
-
-    public async Task ProcessMessageAsync(MyMessage message)
-    {
-        // IO
-        var databaseName = "test";
-        _logger.LogInformation(message.Text);
-        
-        using var session = _client.StartIdempotentTransaction(message, databaseName);
-        var collection = _client.GetDatabase(databaseName).GetCollection<MyMessage>("test.collection");
-
-        await collection.InsertOneAsync(session, message);
     }
 }
