+++ scripts/start-tests.sh
@@ -3,10 +3,9 @@
 # https://www.willianantunes.com/blog/2021/05/production-ready-shell-startup-scripts-the-set-builtin/
 set -eu -o pipefail
 
-CSPROJ_PATH=./tests/Ziggurat.Tests
 REPORTS_FOLDER_PATH=tests-reports
 
-dotnet test $CSPROJ_PATH \
+dotnet test \
     --logger trx \
     --logger "console;verbosity=detailed" \
     --settings "runsettings.xml" \
