+++ src/Ziggurat.MongoDB/MiddlewareOptionsExtensions.cs
@@ -1,0 +1,20 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+using Ziggurat.MongoDB;
+
+namespace Ziggurat;
+
+public static class MiddlewareOptionsExtensions
+{
+    public static void UseMongoDbIdempotency<TMessage>(this MiddlewareOptions<TMessage> options, string mongoDatabaseName)
+        where TMessage : IMessage
+    {
+
+        ZigguratMongoDbOptions.MongoDatabaseName = mongoDatabaseName;
+
+        static void IdempotencySetupAction(IServiceCollection services) =>
+            services
+                .AddScoped<IConsumerMiddleware<TMessage>, IdempotencyMiddleware<TMessage>>();
+
+        options.Extensions.Add(IdempotencySetupAction);
+    }
+}
