+++ src/Ziggurat.MongoDB/Ziggurat.MongoDB.csproj
@@ -1,8 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netstandard2.1</TargetFramework>
-        <LangVersion>9.0</LangVersion>
+        <TargetFrameworks>netstandard2.1;net6.0</TargetFrameworks>
+        <LangVersion>10.0</LangVersion>
     </PropertyGroup>
 
     <ItemGroup>
