+++ src/Ziggurat/Internal/LoggerExtensions.cs
@@ -4,7 +4,7 @@ namespace Ziggurat.Internal
 {
     public static class LoggerExtensions
     {
-        public static void LogMessageExists<T, TMessage>(this ILogger<T> logger, TMessage message)
+        internal static void LogMessageExists<T, TMessage>(this ILogger<T> logger, TMessage message)
             where TMessage : IMessage
             => logger.LogInformation(
                 "Message was processed already. Ignoring {MessageId}:{Type}.", message.MessageId, message.MessageGroup);
