+++ src/Ziggurat/MiddlewareOptions.cs
@@ -1,8 +1,6 @@
 using System;
 using System.Collections.Generic;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
-using Ziggurat.Idempotency;
 
 namespace Ziggurat
 {
@@ -10,17 +8,6 @@ namespace Ziggurat
         where TMessage : IMessage
     {
         internal List<Action<IServiceCollection>> Extensions { get; } = new();
-        public string MongoDatabaseName { get; set; }
-
-        public void UseIdempotency<TContext>() where TContext : DbContext
-        {
-            static void IdempotencySetupAction(IServiceCollection services) =>
-                services.AddScoped<
-                    IConsumerMiddleware<TMessage>,
-                    IdempotencyMiddleware<TMessage, TContext>>();
-
-            Extensions.Add(IdempotencySetupAction);
-        }
 
         public void Use<TMiddleware>()
             where TMiddleware : class, IConsumerMiddleware<TMessage>
